refactor: move all SignUp protocols to a single file
